requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
requests
python-multipart
Pillow
filetype
beautifulsoup4
boto3
botocore
python-jose[cryptography]
passlib[bcrypt]